+++ controller/doser/types.go
@@ -21,11 +21,11 @@ type DosingRegiment struct {
 	Enable   bool          `json:"enable"`
 	Schedule Schedule      `json:"schedule"`
 	Duration time.Duration `json:"duration"`
-	Speed    int           `json:"speed"`
+	Speed    float64       `json:"speed"`
 }
 
 type CalibrationDetails struct {
-	Speed    int           `json:"speed"`
+	Speed    float64       `json:"speed"`
 	Duration time.Duration `json:"duration"`
 }
